test: use int64 for blobberFactory to prefent overflows on 32-bit systems

+++ testutil/blobber.go
@@ -45,7 +45,7 @@ type BlobberFactory struct {
 	rngCache []byte
 }
 
-func NewBlobberFactory(initialRngCacheSize int, seed SeedT) *BlobberFactory {
+func NewBlobberFactory(initialRngCacheSize int64, seed SeedT) *BlobberFactory {
 	res := &BlobberFactory{
 		rngCache: make([]byte, initialRngCacheSize),
 	}
